Change chart.view API from "yaxis" to "show" and "of"
Previously the chart was called by providing a xaxis and a yaxis
property. To better fit to other view methods this is changed to
use "show" and "of" to determine the element to represent on the chart.

One consequence of this change is, that the yaxis can represent only
one element (and not multiple by providing an array).

The xaxis is determined by a ref property (which does not differ much)
from the previously used xaxis property.
The naming of this property is a matter of future discussions.

The previously used ref property is therefore is no longer used to
determine the referenced drive to get the nod on the chart for preciew.
This drive is no automatically identified by
model.inputControlDrives[0].
One limitation is that this only works reliable if only one drive is
used as input (but this limitation is not new...)

If the ref property is not provided, the timer of the model is used.

Other changes are made to make the code more readable.