Put nav and mec logo into seperate functions
The nav is ought to be replacable by third parties.
Therefore the nav itself is going to be one function call.
The respective elements that are affected by this are going
to be replaced aswell.

The logo may be used in different places, therefore it gets a seperate property.
As a bonus: The code is easier to read...